Fix build, cleanup and reorganize
- Remove leftover remnand of mp4parse_fallible dependency
- Move mp4parse::fallible mod to separate file
- Fix clippy warning
